fix feedback from Code Review
